requirements.txt
ddddocr~=1.0.6
psutil~=5.9.1
pywinauto~=0.6.8
selenium~=4.4.0
Pillow~=9.2.0